Update 2018-06-19-missing.md
Could have been that missing values are always at the end : meaning sorted as out of non missing ! It is not the case ! May be it is more consistent but I guess I would have liked the other option. In any case, I only want your text be precise about that.